% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_cuttable}
\alias{is_cuttable}
\title{Determines if a vector should be cut when moedizing}
\usage{
is_cuttable(x, cuts)
}
\description{
Determines if a vector should be cut. Confirms if it inherits a class
for which base::cut is implemented (cut.default uses is.numeric), then checks
if there are more unique values than cuts requested.
}
