% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moeda_function.R
\name{moeda}
\alias{moeda}
\title{Run a MOEDA Analysis}
\usage{
moeda(df, target_var, n_top_vars = 4, cuts = 4, ...)
}
\arguments{
\item{df}{A dataset you wish to analyze, typically a data.frame or a tibble}

\item{target_var}{Unquoted expression or a scalar character representing the
column in the `df` that is your target (aka dependent) variable for analysis}

\item{n_top_vars}{An integer scalar specifying up to how many top variables
you wish to consider for your analysis. By default 4}

\item{cuts}{An integer scalar telling `base::cut()` how many equal width
buckets to create in the top_vars. By default 4}

\item{...}{Arguments passed to `base::cut()`}
}
\value{
The original df as a tibble with joined top features columns that
were cut together with resulting intersections. These additional columns
have `moedized` in their name.
}
\description{
The main function of the package that runs Modeling-Oriented Exploratory
Data Analysis on your data.
}
\details{
Currently, this function performs the following actions:
\enumerate{
  \item A random forest is run on the training set with 80 percent of
  observations
  \item Permutation variable importance is exported from the model and
  reported to the console together with some additional information
  \item Model's performance is assessed on the test set of the remaining
  20 percent of observations and reported to the console
  \item Top variables (their number can be specified with `n_top_vars`
  argument) are discretized using equal widths discretization via
  `base::cut()`. You can select the number of cuts with the `cuts` argument.
  \item An upset plot of the intersections from 4. together with target
  variables is printed.
  \item A `GGally::ggpairs()` plot of top variables is printed.
  \item The function returns the original `df` and joins top features
  columns that were cut together with resulting intersections.
  These additional columns have `moedized` in their name.
}
}
